AD-bind-principal: New abstraction for ActiveDirectory LDAP auth
Most use cases don't require another implementation than the standard
way of binding and searching for a user in Active Directory.

However, sometimes there is a different implementation needed
for creating the principal string for binding a user and the principal
string for searching a user.

The new abstract class allows for users to create a different
implementation according to their needs, based on the given
authentication token (or a subclass of the authentication token).
